+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+  - Nuvlabox-status - Use delete attributes
+  - Job - Use delete-attributes
+  - Nuvlabox - Delete location possible
+
 ## [5.19.1] - 2021-11-10
 
 ### Changed